update: user検索処理の実装修正
