+++ src/app/Http/Requests/Gym/TrainerSearchRequest.php
@@ -26,6 +26,7 @@ class TrainerSearchRequest extends FormRequest implements SearchRequestInterface
     {
         return [
             'user_id' => 'nullable|integer',
+            'user_type' => 'nullable|in:App\Models\Gym,App\Models\Trainer',
             'occupation_id' => 'nullable|integer',
             'area_id' => 'nullable|integer',
             'price' => 'nullable|array',
