+++ src/app/Http/Requests/SearchRequestInterface.php
@@ -4,5 +4,6 @@ namespace App\Http\Requests;
 
 interface SearchRequestInterface
 {
-    
+    public function validated();
+    public function rules();
 }
